put ColorWheel into the utilities::color module
