+++ common/src/utilities.rs
@@ -1,0 +1,1 @@
+pub mod color;
